# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/webhooks/mural": {
      "post": {
        "operationId": "WebhooksController_handleMural",
        "parameters": [
          {
            "name": "x-mural-webhook-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-mural-webhook-timestamp",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mural webhook ingress (account_credited → mark order PAID)",
        "tags": [
          "webhooks"
        ]
      }
    }
  },
  "info": {
    "title": "Mural Marketplace API",
    "description": "Sandbox marketplace backend integrating with Mural Pay staging.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}